Minor codestyle improvements and fixes

+++ src/renderer/rasterizer/rasterizer_renderer.cpp
@@ -9,8 +9,8 @@ void cg::renderer::rasterization_renderer::init()
 	model->load_obj(settings->model_path);
 
 	camera = std::make_shared<cg::world::camera>();
-	camera->set_width(settings->width);
-	camera->set_height(settings->height);
+	camera->set_width(static_cast<float>(settings->width));
+	camera->set_height(static_cast<float>(settings->height));
 	camera->set_position(float3{
 		settings->camera_position[0],
 		settings->camera_position[1],
@@ -23,8 +23,8 @@ void cg::renderer::rasterization_renderer::init()
 	camera->set_z_far(settings->camera_z_far);
 
 	render_target = std::make_shared<cg::resource<cg::unsigned_color>>(settings->width, settings->height);
-	depth_buffer = std::make_shared<cg::resource<float>>(settings->width, settings->height);
-	rasterizer = std::make_shared<cg::renderer::rasterizer<cg::vertex, cg::unsigned_color>>();
+	depth_buffer  = std::make_shared<cg::resource<float>>(settings->width, settings->height);
+	rasterizer    = std::make_shared<cg::renderer::rasterizer<cg::vertex, cg::unsigned_color>>();
 
 	rasterizer->set_render_target(render_target, depth_buffer);
 	rasterizer->set_viewport(settings->width, settings->height);
@@ -58,6 +58,7 @@ void cg::renderer::rasterization_renderer::render()
 	for (size_t shape_id = 0; shape_id < model->get_index_buffers().size(); ++shape_id) {
 		rasterizer->set_vertex_buffer(model->get_vertex_buffers()[shape_id]);
 		rasterizer->set_index_buffer(model->get_index_buffers()[shape_id]);
+
 		rasterizer->draw(model->get_index_buffers()[shape_id]->get_number_of_elements(), 0);
 	}
